refactor: restructure movie service by splitting into modular components for better maintainability and clarity
